Complete API without tests

+++ API/src/main/java/leandroasano/api/Controllers/ProductCategoryController.java
@@ -9,6 +9,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import javax.servlet.http.HttpSession;
 import java.util.List;
 
 @RestController
@@ -17,53 +18,60 @@ public class ProductCategoryController {
     @Autowired
     ProductCategoryService productCategoryService;
 
+    @Autowired
+    HttpSession session;
+
     @PostMapping("/productcategory/add")
     public ResponseEntity addProductCategory(@RequestBody ProductCategory productCategory) throws Exception{
-        try {
-            productCategoryService.createProductCategory(productCategory);
-            System.out.println("Category "+productCategory.getName()+"Created");
-            return new ResponseEntity(HttpStatus.CREATED);
-        } catch (Exception e){
-            throw new Exception("Error adding category");
-        }
+       try {
+           if (session.getAttribute("role").equals("admin")) {
+               productCategoryService.createProductCategory(productCategory);
+               System.out.println("Category " + productCategory.getName() + "Created");
+               return new ResponseEntity(HttpStatus.CREATED);
+           } else {
+               throw new Exception("Error of permissions!");
+           }
+       } catch (Exception e){
+           throw e;
+       }
     }
 
-    @PutMapping("/{idproductcategory}/update")
-    public ResponseEntity updateProductCategory(@PathVariable("idproductcategory") int idprodcat,@RequestBody ProductCategory productCategory) throws Exception {
-        try{
-            productCategoryService.updateProductCategory(idprodcat, productCategory);
+    @PutMapping("/productcategory/{idproductcategory}")
+    public ResponseEntity updateProductCategory(@PathVariable("idproductcategory") int idproductcategory,@RequestBody ProductCategory productCategory) throws Exception {
+        if (session.getAttribute("role").equals("admin")){
+            productCategoryService.updateProductCategory(idproductcategory, productCategory);
             return new ResponseEntity(HttpStatus.OK);
-        }catch (Exception e){
-            throw new Exception("Error in service");
+        } else {
+            throw new Exception("Error of permissions!");
         }
     }
 
     @GetMapping("/productcategory/read/{idprodcat}")
     public ResponseEntity<List<Product>> readProductCategory(@PathVariable("idprodcat") int idprodcat) throws Exception{
-        try {
+        if (session.getAttribute("role").equals("admin")){
             return new ResponseEntity<>(productCategoryService.readProductCategory(idprodcat),HttpStatus.OK);
-        } catch (Exception e){
-            throw new Exception("Error of service");
+        } else {
+            throw new Exception("Error of permissions!");
         }
     }
 
 
     @GetMapping("/productcategory/all")
     public ResponseEntity<List<ProductCategory>> getAllProductsCategroy() throws Exception{
-        try {
+        if (session.getAttribute("role").equals("admin")){
             return new ResponseEntity<>(productCategoryService.getAllProductCategories(),HttpStatus.OK);
-        } catch (Exception e){
-            throw new Exception("Error of service");
+        } else {
+            throw new Exception("Error of permissions!");
         }
     }
 
-    @DeleteMapping("/{idprodcat}/delete")
+    @DeleteMapping("/productcategory/{idprodcat}")
     public ResponseEntity deleteUser(@PathVariable("idprodcat") int idprodcat) throws Exception {
-        try{
+        if (session.getAttribute("role").equals("admin")){
             productCategoryService.deleteProductCategory(idprodcat);
             return new ResponseEntity(HttpStatus.OK);
-        }catch (Exception e){
-            throw new Exception("Error in service");
+        } else {
+            throw new Exception("Error of permissions!");
         }
     }
 }
